Support provider ready event actions, efactor fundAccounts, handle no eth_coinbase
Provide support for actions to be run when blockchain provider is ready. This occurs after provider ready and before the dev accounts are funded. This is to give blockchain plugins an opportunity to create/unlock accounts before account funding occurs.

Refactor `fundAccounts` to make it asynchronous (much easier to read).

Handle case where blockchain does not support `eth_coinbase`, does not return an address, or the address returned is a 0x0 address. In this case, find the account with the most funds and use that at the `from` account when funding development accounts.